refactored structure, made verbose, add setup.py
