+++ pkg_manager/aes.py
@@ -1,4 +1,3 @@
-import hashlib
 from Crypto import Random
 from Crypto.Cipher import AES
 from Crypto.Protocol.KDF import scrypt
