+++ pkg_manager/package_repo.py
@@ -1,4 +1,4 @@
-from package import Package
+from .package import Package
 
 
 class PackageRepo:
