+++ pkg_manager/settings.py
